% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfidfind.R
\name{dfidfind}
\alias{dfidfind}
\title{Finds dataframe ids}
\usage{
dfidfind(resp)
}
\arguments{
\item{resp}{= a list from the output of a dataframes search using refreq}
}
\value{
An atomic vector containing dataframe ids.
}
\description{
This takes a response from a dataframes search then loads the ids into an atomic vector.
}

