% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfreq.R
\name{dfreq}
\alias{dfreq}
\title{Makes a json file from a data frame}
\usage{
dfreq(df, file = "reqobj.json")
}
\arguments{
\item{df}{= dataframe This is the data frame you want to make a request from. If you want to make an array put a list into the dataframe. i.e. df$names <- list("tcga-gene-expression")}

\item{file}{= string This is where you name your file. Remember to end in .json i.e. file = "searchobj.json"}
}
\description{
This will make a json file from a data frame and then write it to your r directory so it can be used in the refreq function.
}

