% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfsearch.R
\name{dfsearch}
\alias{dfsearch}
\title{Sets up a dataframe search request}
\usage{
dfsearch(keyspaceIds, dataframeIds = NULL, unitIds = NULL,
  pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{keyspaceIds}{= vector of strings}

\item{dataframeIds}{= vector of strings}

\item{unitIds}{= vector of strings}

\item{pageSize}{= integer}

\item{pageToken}{= string}
}
\value{
A properly formatted request
}
\description{
formats the request to send through the refreq function
}

