% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genematrix.R
\name{genematrix}
\alias{genematrix}
\title{Converts a dataframe/slice response into a matrix}
\usage{
genematrix(resp)
}
\arguments{
\item{resp}{= a responce from a dataframe/slice request}
}
\value{
A matrix structured with genes as rows and samples as collumns.
}
\description{
Converts a response from a dataframe/slice request into a matrix stuctured with genes as rows and samples as collumns.
}

