% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keysearch.R
\name{keysearch}
\alias{keysearch}
\title{Sets up a keyspaces search request}
\usage{
keysearch(axisNames, keyspaceIds = NULL, names = NULL, keys = NULL,
  pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{axisNames}{= an atomic vector of strings e.g. c("a", "b", "c")}

\item{keyspaceIds}{= an atomic vector of strings e.g. c("a", "b", "c")}

\item{names}{= an atomic vector of strings e.g. c("a", "b", "c")}

\item{keys}{= an atomic vector of strings e.g. c("a", "b", "c")}

\item{pageSize}{= integer}

\item{pageToken}{= string}
}
\value{
A properly formatted request
}
\description{
formats the request to send through the refreq function
}

