% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refreq.R
\name{refreq}
\alias{refreq}
\title{Request from refence server}
\usage{
refreq(url = "http://192.168.99.100:5000/", end = "units/search", req = {
  })
}
\arguments{
\item{url}{= string This is the url you want the data to come from e.g. "http://example.com/api/"}

\item{end}{= string This is the prefex to the above url e.g. "example/search" (This is mostly for convenience sake and is not necisarry if the url section has everything in it)}

\item{req}{= list Input a list containing the search request structured as such: list(axisNames = "sample", keys = c("TCGA-ZQ-A9CR-01A-11R-A39E-31", "TCGA-ZR-A9CJ-01B-11R-A38D-31"))}
}
\value{
a list containig the response
}
\description{
Sends a request in the form of a json file to the framespace reference server and retreaves the content of the response in the form of a list.
}

