% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitssearch.R
\name{unitssearch}
\alias{unitssearch}
\title{Sets up a units search request}
\usage{
unitssearch(ids = NULL, names = NULL, pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{ids}{= an atomic vector of strings e.g. c("a", "b", "c")}

\item{names}{= an atomic vector of strings e.g. c("a", "b", "c")}

\item{pageSize}{= integer}

\item{pageToken}{= string}
}
\value{
A properly formatted request
}
\description{
formats the request to send through the refreq function
}

